Simplify the interface of the ANS coder.
Use only WriteHistograms and WriteTokens, create separate
header for Histogram and Token, which are independent of
ans coding, and make clustering histograms in-place.
